+++ src/module/add_todo.js
@@ -1,0 +1,17 @@
+import todos from './todo';
+
+const addTodos = (e) => {
+  const addTodo = document.querySelector('.add_todo');
+  const todoObject = {
+    description: addTodo.value,
+    completed: 'checked',
+    index: 4,
+  };
+  todos.push(todoObject);
+  localStorage.setItem('todo', JSON.stringify(todos));
+  e.preventDefault();
+  window.location.reload();
+  return todos;
+};
+
+export default addTodos;
