+++ src/module/todo.js
@@ -1,0 +1,3 @@
+const todos = JSON.parse(localStorage.getItem('todo')) || [{ description: '2nd project about to end', completed: '', index: 2 }];
+
+export default todos;
